Various IISNode compatibility changes
IISNode passes an IIS port string through process.env.PORT.  It works fine when passed through to app.listen, but it might not be a number anymore.

Looks like sometimes ctx refuses to show the IP as well, so added some extra error checking around that.

When the private key is supplied through the web.config, it gets extra double quotes around it that need to be trimmed out.